Added <p> and more text in CDATA
I edited the CDATA section to see how it worked. Since I'm fairly new to google gadgets, I'm learning and researching. This is a test to see if the gadget still works wit <p> tags and to see a change in text so I know the site is working correctly.